initial commit of DeCal DigitalOcean VM terraform configuration
Probably not optimal use of Terraform, but this should suffice
while we learn more about the tool. Instantiates a staff VM and
a VM per student, as well as a firewall limiting access to the
UC Berkeley subnet. Next steps are to use Terraform to provision
the servers with our Puppet config.